requirements.txt
PySide6>=6.5.0
requests>=2.28.0
pynput>=1.7.6
mss>=9.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
cryptography>=42.0.0
